fix(wake): keep wakeRestartRecord helper off windows builds
The typed vanished-binary predicate stays shared; the restart-record
helper now uses the same darwin||linux constraint as the type.
