+++ internal/cli/wake_binary_dir_gone_unix.go
@@ -1,0 +1,14 @@
+//go:build darwin || linux
+
+package cli
+
+func wakeRestartRecordBinaryDirGone(record wakeRestartRecord) bool {
+	paths := []string{record.StagePath}
+	if record.BoundImage != nil {
+		paths = append(paths, record.BoundImage.ExecutionPath)
+	}
+	if record.PreviousBoundImage != nil {
+		paths = append(paths, record.PreviousBoundImage.ExecutionPath)
+	}
+	return wakeAnyRecordedPathParentGone(paths)
+}
